feat: refactor SMTC media test by moving utility functions to a new module and updating imports
